test: add backend and frontend vitest suites
