+++ backend/vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    environment: 'node',
+    setupFiles: ['./test/setup-env.ts'],
+    include: ['test/**/*.test.ts']
+  }
+});
